edit(hotkeys): ons for categories + one interface
Make the same interface for adding a location with nothing and Symbol
and make different ons for different categories. Now each category has
its own two listeners. One for adding a categorical location and one for
making the selected location on a given category. This makes it faster
as the listener does not have to loop over all the categories to
determine, what should happen.