edit(locations): Change recipe for locations
new recipe `locationslayer` only manages one layer of the locations in
the image. It is specified with an offset (...,-1,0,+1,...) that
modifies the attributes of the locations and an `annotation_attributes`
dict that decides which attributes will be mapped to a given location.

NOTE:  <kunzaatko martinkunz@email.cz>
This does not yet work with an empty vector of locations because there
will be an empty array of attributes passed and that makes makie to
throw an exception.

TODO:  <kunzaatko martinkunz@email.cz>
This should be made more generic with ability to plot any location
that has a defined look under plots.

TODO:  <kunzaatko martinkunz@email.cz>
It is not yet compatible with the DataInspector