edit/fix(locations): some fixes from tests and structvector
This would be a bad API for the user to need to convert the input into a
structvector... Use it as a lifted observable in the plot recipe for
performance reasons but the user can supply a vector.

Some tests detected problems with changind selected etc. these are fixed
now.